refactor WindowManager
